validar al eliminar producto

+++ proyecto/servicios/proyecto/src/main/java/proyecto/rsc/ProductoRsc.java
@@ -2,6 +2,11 @@ package proyecto.rsc;
 
 import java.util.List;
 
+import javax.persistence.EntityManager;
+
+import org.hibernate.Query;
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -12,7 +17,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import proyecto.ctrl.PedidoCtrl;
 import proyecto.ctrl.ProductoCtrl;
+import proyecto.ent.Pedido;
 import proyecto.ent.Producto;
 
 @CrossOrigin(origins="*")
@@ -23,6 +30,12 @@ public class ProductoRsc {
 	@Autowired
 	private ProductoCtrl productoCtrl;
 	
+	@Autowired
+	private PedidoCtrl pedidoCtrl;
+	
+	@Autowired
+    private EntityManager entityManager;
+	
 	@GetMapping
 	public List<Producto> obtenerTodos() {
 		List<Producto> listaList =  (List<Producto>) productoCtrl.findAll();
@@ -31,6 +44,14 @@ public class ProductoRsc {
 	
 	@PutMapping
 	public Producto guardar(@RequestBody Producto producto) {
+		String error = "";
+		error += Utilidades.validarCampo(producto.getNombre(), "nombre");
+		error += Utilidades.validarCampo(producto.getCostoUnitario(), "costo unitario");
+		error += Utilidades.validarCampo(producto.getDescripcion(), "descripcion");
+		
+		if(!error.isEmpty()) {
+			throw new Error(error);
+		}
 		productoCtrl.save(producto);
 		return producto;
 	}
@@ -41,10 +62,23 @@ public class ProductoRsc {
 		return producto;
 	}
 
+	@SuppressWarnings("unchecked")
 	@DeleteMapping(path = "/{id:[0-9]+}")
 	public Producto borrar(@PathVariable("id") Integer id) {
-		Producto producto = (Producto)productoCtrl.findById(id).get();
-		productoCtrl.deleteById(id);
-		return producto;
+		try {	
+			String sql = "SELECT P.* FROM PEDIDOS P JOIN DET_PRODUCTO DET ON P.ID = DET.PEDIDO_ID WHERE DET.PRODUCTO_ID = " + id;
+			List<Pedido> pedidos = entityManager.createNativeQuery(sql,Pedido.class).getResultList();
+			if( pedidos.size() > 0) {
+				throw new Error("El producto pertenece a uno o más pedidos.");
+			}
+			
+			
+			
+			Producto producto = (Producto)productoCtrl.findById(id).get();
+			productoCtrl.deleteById(id);
+			return producto;
+		} catch (Exception e) {
+			throw new Error(e.getMessage());
+		}
 	}
 }
